Add position and rotation propagation - (Task #5)
Add a simple solution in json_product_child to propagate values from
parent to children and also test this behavior.

The current method results in children only having an absolute position
(and rotation) instead of a relative one to their parent. This should be
reconsidered with usability and the export functionality in mind.

---
Task #5: Implement Import Functionality